Organize and extend data adaptors.
