+++ Numerics/Data/DataFrame_Csv.cs
@@ -179,132 +179,4 @@ namespace Meta.Numerics.Data {
 
     }
 
-    // Type adaptors parse text into objects of a particular type.
-
-    internal abstract class TypeAdaptor
-    {
-        public abstract bool IsParsable(string text);
-
-        public abstract object Parse(string text);
-
-        public abstract DataList CreateStorage(string name, bool nullable);
-
-    }
-
-    internal class DoubleAdaptor : TypeAdaptor
-    {
-        public override bool IsParsable(string text)
-        {
-            double value;
-            bool isParsable = Double.TryParse(text, out value);
-            return (isParsable);
-        }
-
-        public override object Parse(string text)
-        {
-            return (Double.Parse(text));
-        }
-
-        public override DataList CreateStorage (string name, bool nullable) {
-            if (nullable) {
-                return (new DataList<double?>(name));
-            } else {
-                return (new DataList<double>(name));
-            }
-        }
-
-    }
-
-    internal class IntAdaptor : TypeAdaptor
-    {
-        public override bool IsParsable(string text)
-        {
-            int value;
-            bool isParsable = Int32.TryParse(text, out value);
-            return (isParsable);
-        }
-
-        public override object Parse(string text)
-        {
-            return (Int32.Parse(text));
-        }
-
-        public override DataList CreateStorage (string name, bool nullable) {
-            if (nullable) {
-                return (new DataList<int?>(name));
-            } else {
-                return (new DataList<int>(name));
-            }
-        }
-    }
-
-    internal class DateTimeAdaptor : TypeAdaptor
-    {
-
-        public override bool IsParsable(string text)
-        {
-            DateTime value;
-            bool isParsable = DateTime.TryParse(text, out value);
-            return (isParsable);
-        }
-
-        public override object Parse(string text)
-        {
-            return (DateTime.Parse(text));
-        }
-
-        public override DataList CreateStorage(string name, bool nullable)
-        {
-            if (nullable) {
-                return (new DataList<DateTime?>(name));
-            } else {
-                return (new DataList<DateTime>(name));
-            }
-        }
-
-    }
-
-    internal class TimeSpanAdaptor : TypeAdaptor {
-
-        public override bool IsParsable (string text) {
-            Debug.Assert(text != null);
-            // TimeSpan's parse will accept pure ints or doubles as timespans.
-            // Since we don't want a coulumn of ints or doubles to become timespans,
-            // we require a colon before we even try.
-            if (!text.Contains(":")) return (false);
-            TimeSpan value;
-            bool isParsalbe = TimeSpan.TryParse(text, out value);
-            return (isParsalbe);
-        }
-
-        public override object Parse (string text) {
-            TimeSpan value = TimeSpan.Parse(text);
-            return (value);
-        }
-
-        public override DataList CreateStorage (string name, bool nullable) {
-            if (nullable) {
-                return (new DataList<TimeSpan?>(name));
-            } else {
-                return (new DataList<TimeSpan>(name));
-            }
-        }
-
-    }
-
-    // This type is used to keep track of which types are compatible with the text
-    // representations that have been read in for a column.
-
-    internal class DataAdaptor
-    {
-        private static TypeAdaptor[] knownTypes = new TypeAdaptor[] {
-            new DateTimeAdaptor(), /* new TimeSpanAdaptor(), */ new IntAdaptor(), new DoubleAdaptor()
-        };
-
-        public bool IsNullable = false;
-
-        internal LinkedList<TypeAdaptor> TypeCandidates = new LinkedList<TypeAdaptor>(knownTypes);
-
-    }
-
 }
